Improved plotting script
